{{cookiecutter.project_name}}/requirements.txt
pathlib2
wheel
protobuf>=3.0.0
googleapis-common-protos>=1.3.1
